#795 Move some logic to team utils
