+++ src/backend/src/utils/notifications.utils.ts
@@ -1,6 +1,7 @@
 import { Team, Task as Prisma_Task, WBS_Element } from '@prisma/client';
 import { UserWithSettings } from './auth.utils';
 import { HttpException } from './errors.utils';
+import { getTeamsFromUsers } from './teams.utils';
 
 export type UserWithTeams = UserWithSettings & {
   teamAsHead: Team | null;
@@ -28,7 +29,7 @@ export const getTeamFromTaskAssignees = (users: UserWithTeams[]): Team => {
   const usersTeams = getTeamsFromUsers(users);
 
   firstUsersTeams: for (const team of usersTeams[0]) {
-    for (let i = 0; i < usersTeams.length; i++) {
+    for (let i = 1; i < usersTeams.length; i++) {
       // If the team is not found in the current user's teams, continue to the next team
       if (!usersTeams[i].some((t) => t.teamId === team.teamId)) continue firstUsersTeams;
     }
@@ -36,21 +37,6 @@ export const getTeamFromTaskAssignees = (users: UserWithTeams[]): Team => {
   }
 
   throw new HttpException(400, 'All of the users do not share a team!');
-};
-
-/**
- * Gets the teams from a list of users
- * @param users the users to get the teams from
- * @returns an array of the teams each user is in
- */
-export const getTeamsFromUsers = (users: UserWithTeams[]): Team[][] => {
-  return users.map((user) => {
-    const teams = [];
-    if (user.teamAsHead) teams.push(user.teamAsHead);
-    if (user.teamsAsLead) teams.push(...user.teamsAsLead);
-    if (user.teamsAsMember) teams.push(...user.teamsAsMember);
-    return teams;
-  });
 };
 
 /**
