+++ src/backend/src/utils/teams.utils.ts
@@ -1,4 +1,5 @@
-import { Prisma, User } from '@prisma/client';
+import { Prisma, User, Team } from '@prisma/client';
+import { UserWithTeams } from './notifications.utils';
 
 const teamQueryArgsMembersOnly = Prisma.validator<Prisma.TeamArgs>()({
   include: {
@@ -49,4 +50,19 @@ export const areUsersPartOfTeams = (teams: Prisma.TeamGetPayload<typeof teamQuer
  */
 export const isUserPartOfTeams = (teams: Prisma.TeamGetPayload<typeof teamQueryArgsMembersOnly>[], user: User) => {
   return teams.some((team) => isUserOnTeam(team, user));
+};
+
+/**
+ * Gets the teams from a list of users
+ * @param users the users to get the teams from
+ * @returns an array of the teams each user is in
+ */
+export const getTeamsFromUsers = (users: UserWithTeams[]): Team[][] => {
+  return users.map((user) => {
+    const teams = [];
+    if (user.teamAsHead) teams.push(user.teamAsHead);
+    if (user.teamsAsLead) teams.push(...user.teamsAsLead);
+    if (user.teamsAsMember) teams.push(...user.teamsAsMember);
+    return teams;
+  });
 };
